@@ -22,9 +22,13 @@
     ],
     "require": {
         "ext-curl": "*",
+        "ext-dom": "*",
         "ext-mbstring": "*",
         "ext-json": "*",
-        "php": ">=7.2.0"
+        "php": ">=8.0"
+    },
+    "require-dev": {
+        "phpunit/phpunit": "^9"
     },
     "autoload": {
         "psr-4": {
