@@ -8,60 +8,64 @@
  */
 namespace Base\Core;
 
-use Base\Core\Response;
-
 /**
  * Class Call holds controller, method and list of arguments.
- * @package Base\Core
  */
 class Call
 {
     /**
      * Controller.
-     * @var Controller
+     * @var BaseController
      */
-    protected $controller;
+    protected BaseController $controller;
 
     /**
      * Method name.
      * @var string
      */
-    protected $methodName;
+    protected string $methodName;
+
+    /**
+     * @var mixed
+     */
+    protected mixed $ctx;
 
     /**
      * Parameters.
      * @var array
      */
-    protected $params;
+    protected array $params;
 
     /**
      * Call constructor.
-     * @param Controller $controller
+     * @param BaseController $controller
      * @param string $methodName
+     * @param mixed $ctx
      * @param array $params
      */
-    public function __construct(Controller $controller, string $methodName, array $params = [])
+    public function __construct(BaseController $controller, string $methodName, mixed $ctx, array $params = [])
     {
         $this->controller = $controller;
         $this->methodName = $methodName;
+        $this->ctx = $ctx;
         $this->params = $params;
     }
 
     /**
      * Gets parameter.
      * @param string $name
-     * @return mixed|null
+     * @return mixed
      */
-    public function param(string $name)
+    public function param(string $name): mixed
     {
-        return isset($this->params[$name]) ? $this->params[$name] : null;
+        return $this->params[$name] ?? null;
     }
 
     /**
      * Gets controller.
-     * @return Controller
+     * @return BaseController
      */
-    public function controller(): Controller
+    public function controller(): BaseController
     {
         return $this->controller;
     }
@@ -72,16 +76,22 @@ class Call
      */
     public function call(): Response
     {
+        // Parameters
+        $params = [$this->ctx];
+        foreach ($this->params as $param) {
+            $params[] = $param;
+        }
+
         // Check whether method name contains multiple method names in format "method1<method2<method3".
         $methods = explode("<", $this->methodName);
         if (!is_array($methods) || count($methods) <= 1) {
             // In case of single method in callback just call this method as provided.
-            return call_user_func_array([$this->controller, $this->methodName], $this->params);
+            return call_user_func_array([$this->controller, $this->methodName], $params);
         }
 
         // In case of multiple method names call first one, but pass new callback as argument instead of params.
         $methodName = array_shift($methods);
-        $callback = new Call($this->controller, implode("<", $methods), $this->params);
-        return call_user_func_array([$this->controller, $methodName], ["callback" => $callback]);
+        $callback = new Call($this->controller, implode("<", $methods), $this->ctx, $this->params);
+        return call_user_func([$this->controller, $methodName], $this->ctx, $callback);
     }
 }
