@@ -1,7 +1,7 @@
 <?php
 /**
  * @project BasePHP Core
- * @file View.php created by Ariel Bogdziewicz on 26/09/2018
+ * @file BaseView.php created by Ariel Bogdziewicz on 26/09/2018
  * @author Ariel Bogdziewicz
  * @copyright Copyright © 2018 Ariel Bogdziewicz. All rights reserved.
  * @license MIT
@@ -12,20 +12,12 @@ use Base\Data\Data;
 
 /**
  * Base class for all views.
- * @package Base\Core
  */
-abstract class View
+abstract class BaseView
 {
-    /**
-     * View constructor.
-     */
-    public function __construct()
-    {
-    }
-
     /**
      * Returns data to display.
      * @return Data
      */
-    abstract function get(): Data;
+    abstract public function get(): Data;
 }
